requirements.txt
pbr>=0.6,!=0.7,<1.0
python-cinderclient
python-novaclient
oslo.config
oslo.utils
PrettyTable>=0.7,<0.8
